+++ infrastructure/modules/iam/variables.tf
@@ -1,0 +1,10 @@
+variable "project_name" {
+  type = string
+}
+variable "s3_bucket_arn" {
+  type = string
+}
+variable "tags" {
+  type    = map(string)
+  default = {}
+}