+++ infrastructure/modules/lambda/outputs.tf
@@ -1,0 +1,3 @@
+output "function_arn" { value = aws_lambda_function.health_check.arn }
+output "function_name" { value = aws_lambda_function.health_check.function_name }
+output "schedule_rule" { value = aws_cloudwatch_event_rule.every_5min.name }