+++ infrastructure/modules/security/variables.tf
@@ -1,0 +1,18 @@
+variable "project_name" {
+  type = string
+}
+variable "vpc_id" {
+  type = string
+}
+variable "app_port" {
+  type    = number
+  default = 5000
+}
+variable "ssh_cidr" {
+  type    = string
+  default = "0.0.0.0/0"
+}
+variable "tags" {
+  type    = map(string)
+  default = {}
+}