+++ infrastructure/modules/vpc/variables.tf
@@ -1,0 +1,11 @@
+variable "project_name" {
+  type = string
+}
+variable "vpc_cidr" {
+  type    = string
+  default = "10.0.0.0/16"
+}
+variable "tags" {
+  type    = map(string)
+  default = {}
+}